Add new feature:     * Polish the structure of the directory, add a folder named Utils     * Add uick operations, able to load saved thetas and pick it to the dashboard     * Add a new type to the DOF, named DofFreeType

+++ project/utils/csv.h
@@ -190,7 +190,7 @@ public:
         std::memcpy(buffer, str, (size_t)to_copy_byte_count);
         remaining_byte_count -= to_copy_byte_count;
         str += to_copy_byte_count;
-        return to_copy_byte_count;
+        return (int)to_copy_byte_count;
     }
 
     ~NonOwningStringByteSource() {}
